Move `Attributes` and `GearSet` out of `Character`
