+++ src/main/kotlin/profile/ProfileCharacter.kt
@@ -19,6 +19,4 @@ data class ProfileCharacter(
     val dc: String,
     val title: String?,
     val town: Town,
-    val attributes: Attributes?,
-    val gearSet: ProfileGearSet?
 )
