+++ src/test/kotlin/AttributesTest.kt
@@ -1,15 +1,15 @@
-import cloud.drakon.ktlodestone.Character
+import cloud.drakon.ktlodestone.Attributes
 import cloud.drakon.ktlodestone.exception.CharacterNotFoundException
 import kotlin.test.Test
 import kotlinx.coroutines.runBlocking
-import org.junit.jupiter.api.Assertions.assertDoesNotThrow
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.assertThrows
 
-class CharacterTest {
-    @Test fun getCharacter() {
-        assertDoesNotThrow {
+class AttributesTest {
+    @Test fun getAttributes() {
+        Assertions.assertDoesNotThrow {
             return@assertDoesNotThrow runBlocking {
-                println(Character.getCharacter(27545492))
+                println(Attributes.getAttributes(27545492))
             }
         }
     }
@@ -17,7 +17,7 @@ class CharacterTest {
     @Test fun getInvalidCharacter() {
         assertThrows<CharacterNotFoundException> {
             return@assertThrows runBlocking {
-                println(Character.getCharacter(0))
+                println(Attributes.getAttributes(0))
             }
         }
     }
